backend/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.7.2
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
blis==0.7.11
catalogue==2.0.10
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.8
Cython==0.29.36
datasets==3.1.0
DAWG-Python==0.7.2
DAWG2-Python==0.9.0
diagrams==0.21.1
dill==0.3.8
Django==4.2.20
django-ckeditor-5==0.2.18
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.3
django-js-asset==2.2.0
django-rest-passwordreset==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
faiss-cpu==1.8.0.post1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
graphviz==0.19.2
gunicorn==23.0.0
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.4.1
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.1
nltk==3.9.1
numpy==1.24.3
packaging==25.0
pandas==2.0.3
pillow==10.4.0
preshed==3.0.9
propcache==0.2.0
psutil==7.0.0
psycopg==3.2.6
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pygraphviz==1.11
PyJWT==2.8.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy2-dicts-uk==2.4.1.1.1460299261
pymorphy3==2.0.3
pymorphy3-dicts-ru==2.4.417150.4580142
pymorphy3-dicts-uk==2.4.1.1.1663094765
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sentencepiece==0.2.0
six==1.17.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
sympy==1.13.3
thinc==8.2.4
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
transformers==4.46.3
typer==0.9.4
typing_extensions==4.13.0
tzdata==2025.2
uk-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/uk_core_news_sm-3.7.0/uk_core_news_sm-3.7.0-py3-none-any.whl#sha256=782ab2dad21895039aaa77f721acc0a5be383d38447d8971e88750393373927a
urllib3==2.2.3
wasabi==1.1.3
weasel==0.3.4
whitenoise==6.7.0
xxhash==3.5.0
yarl==1.15.2
